requirements.txt
Pillow==5.0.0
PyYAML==3.12
certifi==2017.11.5
chardet==3.0.4
idna==2.6
requests==2.18.4
rtmbot==0.4.0
six==1.11.0
slackclient==1.1.0
urllib3==1.22
websocket-client==0.46.0
zbarlight==1.2
